Adding service layer and controller

+++ Backend/SchoolOPS/SchoolOPS/src/main/java/com/schoolOps/SchoolOPS/repository/AttendenceRepository.java
@@ -3,6 +3,6 @@ package com.schoolOps.SchoolOPS.repository;
 import com.schoolOps.SchoolOPS.entity.Attendence;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttendenceRepository extends JpaRepository<Attendence, Integer> {
+public interface AttendenceRepository extends JpaRepository<Attendence, Long> {
 
 }
