+++ Backend/SchoolOPS/SchoolOPS/src/main/java/com/schoolOps/SchoolOPS/repository/ClassWorkRepository.java
@@ -3,6 +3,6 @@ package com.schoolOps.SchoolOPS.repository;
 import com.schoolOps.SchoolOPS.entity.ClassWork;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassWorkRepository extends JpaRepository<ClassWork, Integer> {
+public interface ClassWorkRepository extends JpaRepository<ClassWork, Long> {
 
 }
