+++ Backend/SchoolOPS/SchoolOPS/src/main/java/com/schoolOps/SchoolOPS/repository/EnqueryRepository.java
@@ -3,6 +3,6 @@ package com.schoolOps.SchoolOPS.repository;
 import com.schoolOps.SchoolOPS.entity.Enquery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnqueryRepository extends JpaRepository<Enquery, Integer> {
+public interface EnqueryRepository extends JpaRepository<Enquery, Long> {
 
 }
