+++ Backend/SchoolOPS/SchoolOPS/src/main/java/com/schoolOps/SchoolOPS/repository/TeacherRepository.java
@@ -3,6 +3,6 @@ package com.schoolOps.SchoolOPS.repository;
 import com.schoolOps.SchoolOPS.entity.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
+public interface TeacherRepository extends JpaRepository<Teacher, Long> {
 
 }
