+++ Backend/SchoolOPS/SchoolOPS/src/main/java/com/schoolOps/SchoolOPS/repository/WorkRepository.java
@@ -3,6 +3,6 @@ package com.schoolOps.SchoolOPS.repository;
 import com.schoolOps.SchoolOPS.entity.Work;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkRepository extends JpaRepository<Work, Integer> {
+public interface WorkRepository extends JpaRepository<Work, Long> {
 
 }
